third_party/WebKit/LayoutTests/fast/events/drag-nested-eventSender-use.html

<!DOCTYPE html>
<html>
<body>
<script src="../../resources/js-test.js"></script>
<script>
description("Test that unsupported nested use of eventSender.beginDragWithFiles() is handled without crashing.");

if (window.testRunner)
    testRunner.dumpAsText();

function runTest()
{
    eventSender.beginDragWithFiles(["resources/empty.html"]);
    eventSender.beginDragWithFiles(["resources/empty.html"]);
    eventSender.mouseMoveTo(10, 10);
    eventSender.mouseUp();
}

shouldThrow("runTest()");
</script>
</body>
</html>
